% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_als.R
\name{als_substructure}
\alias{als_substructure}
\title{Assign New Classes to Each ALS Element}
\usage{
als_substructure(x)
}
\arguments{
\item{x}{an ALS object.}
}
\value{

}
\description{
Assign New Classes to Each ALS Element
}
