% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_strings}
\alias{clean_strings}
\title{Standardized string formatting.}
\usage{
clean_strings(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
A character vector.
}
\description{
Standardized string formatting.
}
