% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_class.R
\name{correct_class}
\alias{correct_class}
\title{Correct the Column Classes}
\usage{
correct_class(x, schema)
}
\arguments{
\item{x}{a vector}

\item{schema}{a pointblank schema that supplies the correct class.}
}
\value{
a data frame.
}
\description{
Correct the Column Classes
}
