% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_json}
\alias{export_json}
\title{Export as JSON}
\usage{
export_json(x, path, filename)
}
\arguments{
\item{x}{an object to be written as a JSON file.}

\item{path}{Where to write the file.}

\item{filename}{The name of the file to be written.}
}
\value{
Write a JSON Object.
}
\description{
Export as JSON
}
