% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_als.R
\name{flatten_als}
\alias{flatten_als}
\title{Flatten ALS Objects into a Single Data Frame.}
\usage{
flatten_als(x)
}
\arguments{
\item{x}{an ALS object.}
}
\value{
a data frame.
}
\description{
Flatten ALS Objects into a Single Data Frame.
}
