% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_section.R
\name{gen_section}
\alias{gen_section}
\title{Generate a New R Markdown Tabset Section}
\usage{
gen_section(x, header)
}
\arguments{
\item{x}{a list object.}

\item{header}{a character string to be used to name the section.}
}
\value{
output to be rendered in R Markdown document.
}
\description{
Generate a New R Markdown Tabset Section
}
