% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_section.R
\name{gen_subsection}
\alias{gen_subsection}
\title{Generate a Subsection}
\usage{
gen_subsection(x, section, header)
}
\arguments{
\item{x}{a list.}

\item{section}{a name of the list item of interest.}

\item{header}{a string to be used as a label for the section.}
}
\value{
an output to be rendered in markdown.
}
\description{
Generate a Subsection
}
