% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_val_report.R
\name{gen_val_report}
\alias{gen_val_report}
\title{Generate a Validation Report}
\usage{
gen_val_report(x, output_dir, filename, ...)
}
\arguments{
\item{x}{a data object.}

\item{output_dir}{a file path were the rendered report will be stored.}

\item{filename}{a character string to include in file name.}

\item{...}{additional arguments.}
}
\value{

}
\description{
Generate a Validation Report
}
