% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_element}
\alias{get_element}
\title{Get a List Element}
\usage{
get_element(list, pattern, clean_names = FALSE)
}
\arguments{
\item{list}{a named list object.}

\item{pattern}{a regular expression to be matched to a character string.}

\item{clean_names}{a logical indicating if column names should be cleaned.}
}
\value{
a single element of a list.
}
\description{
Get a List Element
}
\examples{
get_element(list = list(first_element = 1:3,
 second_element = 4:6),
 pattern = "second")
}
