% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_val_report.R
\name{get_html_report}
\alias{get_html_report}
\title{Import HTML Report}
\usage{
get_html_report(output_dir, filename)
}
\arguments{
\item{output_dir}{a file path were the rendered report will be stored.}

\item{filename}{a character string to include in file name.}
}
\value{
a character string representing an html document.
}
\description{
Import HTML Report
}
