% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schema.R
\name{get_schema}
\alias{get_schema}
\title{Get the Appropriate Schema Object based on Supplied Object Class}
\usage{
get_schema(x)
}
\arguments{
\item{x}{a pointblank schema object.}
}
\value{
a pointblank schema object.
}
\description{
Get the Appropriate Schema Object based on Supplied Object Class
}
