% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_schema_class}
\alias{get_schema_class}
\title{Get Schema Class}
\usage{
get_schema_class(x, class_type)
}
\arguments{
\item{x}{a pointblank schema object.}

\item{class_type}{a character string representing the class of interest.}
}
\value{
a characer vector
}
\description{
Get Schema Class
}
