% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zip.R
\name{get_zip_content}
\alias{get_zip_content}
\title{Get Zip File Content Information}
\usage{
get_zip_content(.zip_path)
}
\arguments{
\item{.zip_path}{a file path to a zip file.}
}
\value{
a data frame containing information about the zip file.
}
\description{
Get Zip File Content Information
}
