% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{list_to_df.ALS}
\alias{list_to_df.ALS}
\title{Flatten ALS Objects into a Single Data Frame.}
\usage{
\method{list_to_df}{ALS}(x)
}
\arguments{
\item{x}{an ALS object.}
}
\value{
a data frame.
}
\description{
Flatten ALS Objects into a Single Data Frame.
}
