% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert a list to a data frame based on special class}
\usage{
list_to_df(x)
}
\arguments{
\item{x}{a list.}
}
\value{
a data frame.
}
\description{
Convert a list to a data frame based on special class
}
