% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest_sdg_ssc}
\alias{nest_sdg_ssc}
\title{Nest a Data Frame by SDG and sys_sample_code}
\usage{
nest_sdg_ssc(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
a list.
}
\description{
Nest a Data Frame by SDG and sys_sample_code
}
