% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_als.R
\name{read_als}
\alias{read_als}
\title{Import ALS Data}
\usage{
read_als(zip_path)
}
\arguments{
\item{zip_path}{a file path to an ALS EDD provided in a zip file.}
}
\value{
a list.
}
\description{
Import ALS Data
}
