% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zip.R
\name{read_htm}
\alias{read_htm}
\title{Read and Clean HTM Strings}
\usage{
read_htm(.zip_path, .file)
}
\arguments{
\item{.zip_path}{a file path to a zip file.}

\item{.file}{the name of a single file contained within a zip file.}
}
\value{
a clean htm string.
}
\description{
Read and Clean HTM Strings
}
