% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zip.R
\name{read_zip}
\alias{read_zip}
\title{Read in Files Contained in a Zip File}
\usage{
read_zip(.zip_path, ...)
}
\arguments{
\item{.zip_path}{a file path to a zip file.}

\item{...}{pass arguements on to read.csv or read.table.}
}
\value{
A list of objects imported from a zip file.
}
\description{
Read in Files Contained in a Zip File
}
