% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_validation.R
\name{run_validation}
\alias{run_validation}
\title{Run Validation Process (API Use Only)}
\usage{
run_validation(input_path)
}
\arguments{
\item{input_path}{The file path for the data to be imported into this script.}
}
\value{
a JSON object.
}
\description{
Run Validation Process (API Use Only)
}
