% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentry-package.R
\docType{package}
\name{Sentry-package}
\alias{Sentry}
\alias{Sentry-package}
\title{Sentry: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BWAM/sentry}
  \item Report bugs at \url{https://github.com/BWAM/sentry/issues}
}

}
\author{
\strong{Maintainer}: Zachary Smith \email{zachary.smith2@dec.ny.gov}

}
\keyword{internal}
