% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_class.R
\name{switch_class}
\alias{switch_class}
\title{Switch Class Based on Expected Class}
\usage{
switch_class(x, expected_class)
}
\arguments{
\item{x}{a vector}

\item{expected_class}{a character string.}
}
\value{

}
\description{
Switch Class Based on Expected Class
}
