% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim_underscore}
\alias{trim_underscore}
\title{Trim Leading and Trailing Underscores}
\usage{
trim_underscore(x)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Trim Leading and Trailing Underscores
}
