% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{underscore}
\alias{underscore}
\title{Replace non-numeric or non-character values with "_".}
\usage{
underscore(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
A character vector.
}
\description{
Replace non-numeric or non-character values with "_".
}
