% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_expected_values.R
\name{val_expected_values}
\alias{val_expected_values}
\title{Validate Expected Values in Columns}
\usage{
val_expected_values(x, actions)
}
\arguments{
\item{actions}{}
}
\value{
A pointblank interrogation.
}
\description{
Validate Expected Values in Columns
}
