% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_names.R
\name{val_names}
\alias{val_names}
\title{Validated Column Names}
\usage{
val_names(x, expected, ...)
}
\arguments{
\item{x}{a data frame of provided data.}

\item{expected}{a pointblank schema of expected data.}

\item{...}{Arguments to be passed on to pointblank::col_exists.}
}
\value{
pointblank interrogated data.
}
\description{
Validated Column Names
}
