% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_summary.R
\name{val_summary}
\alias{val_summary}
\title{Validation Summary}
\usage{
val_summary(
  kvp_element,
  kvp_value,
  email_address,
  filename,
  error_summary,
  report,
  status
)
}
\arguments{
\item{kvp_element}{The name of the json element that will be used to query validation reports.}

\item{kvp_value}{The unique value that will be stored in \code{kvp_element} that will be used to query validation report.}

\item{email_address}{a character vector with at least one email address. The provided emails must be address to "dec.ny.gov" accounts. \%>\%}

\item{filename}{a string representing the name of the file under validation.}

\item{error_summary}{a character vector where each element represents a summary of errors or lack there of.}

\item{report}{a validation report.}

\item{status}{a character string indicating if the data deliverable passed ("pass") or failed ("fail") validation.}
}
\value{
a list
}
\description{
Validation Summary
}
