% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Workflow}
\usage{
workflow(x, filename, gen_report = TRUE)
}
\arguments{
\item{x}{a list with a secondary class, such as ALS.}

\item{filename}{a string.}

\item{gen_report}{a logical indicating if a report should be generated.}
}
\value{
a list.
}
\description{
Workflow
}
