% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTemplate-package.R
\docType{package}
\name{Sentry-package}
\alias{Sentry}
\alias{Sentry-package}
\title{Sentry: Data Verification}
\description{
This package provides a toolset to verify that data has been provided in an expected format and conforms to expected standards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BWAM/sentry}
  \item Report bugs at \url{https://github.com/BWAM/sentry/issues}
}

}
\author{
\strong{Maintainer}: Zachary Smith \email{zachary.smith2@dec.ny.gov}

}
\keyword{internal}
