% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Workflow}
\usage{
workflow(x, filepath, filename)
}
\arguments{
\item{x}{a list with a secondary class, such as ALS.}

\item{filepath}{a file path.}

\item{filename}{a string.}
}
\value{
a list.
}
\description{
Workflow
}
